🐛 fix(agentes): render client-only (ssr:false) para eliminar React #418
Root cause del crash de hidratación en /agentes reportado por QA (5-ago):
la vista es 100% client-data-driven (sesiones Zustand, métricas api-ia, tema
claro/oscuro) pero SSR-eaba completa → el HTML del servidor no coincidía con el
1er render del cliente (mismatch de tema = "pantalla dividida negro/blanco") →
hidratación abortada → los useEffect de fetch nunca disparaban ("cero requests").

El guard `mounted` a nivel de página (#265) fue insuficiente: el mismatch viene
del árbol SSR de la ruta, no solo del cuerpo. Solución robusta e idéntica al
patrón ya usado en /asistente: NO renderizar la ruta en el servidor.

- page.tsx: wrapper thin que carga la vista con dynamic(..., { ssr: false }).
- AgentsCoworkView.tsx: la vista completa (movida desde page.tsx, sin cambios de
  lógica) — git detecta rename.
